Use flask routes instead of controllers
Also moved all the non-route methods from the controller file into more sensible places (hopefully)
