+++ ckanext/ldap/logic/auth/update.py
@@ -4,9 +4,9 @@
 # This file is part of ckanext-ldap
 # Created by the Natural History Museum in London, UK
 
-from ckanext.ldap.controllers.user import _find_ldap_user
 from ckanext.ldap.model.ldap_user import LdapUser
 from ckanext.ldap.plugin import config
+from ckanext.ldap.lib.search import find_ldap_user
 
 from ckan.logic import auth
 from ckan.plugins import toolkit
@@ -18,8 +18,8 @@ def user_update(next_auth, context, data_dict):
     '''Ensure LDAP users cannot be edited, and name clash with ldap users
 
     :param next_auth: the next auth function in the chain
-    :param context: 
-    :param data_dict: 
+    :param context:
+    :param data_dict:
 
     '''
     user_obj = None
@@ -36,7 +36,7 @@ def user_update(next_auth, context, data_dict):
             }
     # Prevent name clashes!
     if u'name' in data_dict and user_obj and user_obj.name != data_dict[u'name']:
-        ldap_user_dict = _find_ldap_user(data_dict[u'name'])
+        ldap_user_dict = find_ldap_user(data_dict[u'name'])
         if ldap_user_dict:
             if len(user_obj.ldap_user) == 0 or user_obj.ldap_user[0].ldap_id != \
                     ldap_user_dict[u'ldap_id']:
