+++ ckanext/ldap/routes/__init__.py
@@ -1,0 +1,9 @@
+# !/usr/bin/env python
+# encoding: utf-8
+#
+# This file is part of ckanext-ldap
+# Created by the Natural History Museum in London, UK
+
+from . import user
+
+blueprints = [user.blueprint]
